Move scenario-based dataclasses.
